# interp - Prefer broadcast over reindex when possible