--- a/xarray/core/dataset.py
+++ b/xarray/core/dataset.py
@@ -3855,7 +3855,9 @@ class Dataset(
 
                 if matching_dims and all(var.sizes[d] == 1 for d in matching_dims):
                     # Broadcastable, can be handled quickly without reindex:
-                    to_broadcast = (var.squeeze(),) + tuple(dest for index, dest in use_indexers.values())
+                    to_broadcast = (var.squeeze(),) + tuple(
+                        dest for index, dest in use_indexers.values()
+                    )
                     variables[name] = broadcast_variables(*to_broadcast)[0]
                 elif matching_dims:
                     # For types that we do not understand do stepwise