--- a/xarray/core/dataset.py
+++ b/xarray/core/dataset.py
@@ -3850,16 +3850,14 @@ class Dataset(
                 # For normal number types do the interpolation:
                 var_indexers = {k: v for k, v in use_indexers.items() if k in var.dims}
                 variables[name] = missing.interp(var, var_indexers, method, **kwargs)
-            elif dtype_kind in "ObU":
-                matching_dims = use_indexers.keys() & var.dims
-
-                # if matching_dims and all(var.sizes[d] == 1 for d in matching_dims):
-                # Broadcastable, can be handled quickly without reindex:
-                # to_broadcast = (var.squeeze(),) + tuple(
-                # dest for index, dest in use_indexers.values()
-                # )
-                # variables[name] = broadcast_variables(*to_broadcast)[0]
-                if matching_dims:
+            elif dtype_kind in "ObU" and use_indexers.keys() & var.dims:
+                if all(var.sizes[d] == 1 for d in (use_indexers.keys() & var.dims)):
+                    # Broadcastable, can be handled quickly without reindex:
+                    to_broadcast = (var.copy().squeeze(),) + tuple(
+                        dest for index, dest in use_indexers.values()
+                    )
+                    variables[name] = broadcast_variables(*to_broadcast)[0]
+                else:
                     # For types that we do not understand do stepwise
                     # interpolation to avoid modifying the elements.
                     # reindex the variable instead because it supports