When a variable is a scalar it is faster to broadcast instead of using reindex. Use that when doing dataset interpolation.

- [x] Closes #10054
- [x] Tests added
- [x] User visible changes (including notable bug fixes) are documented in `whats-new.rst`

```python
import numpy as np
import dask.array as da

import xarray as xr

ds = xr.Dataset(
    data_vars={
        "variable_name": (
            "time",
            da.from_array(np.array(["test"], dtype=str), chunks=(1,)),
        )
    },
    coords={"time": ("time", np.array([0]))},
)

%timeit ds.interp(time=np.linspace(0, 10, 50))
%timeit ds.interp(time=np.linspace(0, 10, 100))
%timeit ds.interp(time=np.linspace(0, 10, 1000))
```

Main:
```
16.8 ms ± 620 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)
27.5 ms ± 2.62 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
214 ms ± 7.5 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

PR:
```
5.3 ms ± 138 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)
5.29 ms ± 87 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)
5.29 ms ± 188 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```